+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: alrobert <alrobert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/26 14:39:38 by alrobert          #+#    #+#             */
-/*   Updated: 2022/09/27 14:14:25 by alrobert         ###   ########.fr       */
+/*   Updated: 2022/10/26 09:35:34 by alrobert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned int	i;
 	char			*str;
