+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: alrobert <alrobert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/28 10:12:15 by alrobert          #+#    #+#             */
-/*   Updated: 2022/10/17 18:33:34 by alrobert         ###   ########.fr       */
+/*   Updated: 2022/10/26 10:06:52 by alrobert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	size_t				i;
-	size_t				len_src;
+	size_t	i;
+	size_t	len_src;
 
-	len_src = (size_t)ft_strlen((char *)src);
+	len_src = ft_strlen((char *)src);
 	if (size == 0)
 		return (len_src);
 	i = 0;
